fix(oneseo): 원서 생성시 접수번호 할당 되도록 변경
